Added suspense to model selection provider
